Test proxying an RPC stub all the way to Durable Objects.
In particular, the DO keeps a dup() of the stub and uses it later, after the original subscription call has returned. This requires the `rpc_params_dup_stubs` compat flag, which is new in yesterday's workerd release.

Since wrangler is not updated yet, this requires a package version override for now. Once wrangler has been updated, I'll remove the override and then merge this.

DO NOT MERGE until the override can be removed.

Fixes #110.